requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Caching==2.1.0
Werkzeug==3.0.1
PyMongo==4.6.2
python-dotenv==1.0.0
